--- __init__.py
@@ -1,1 +1,0 @@
-from infotopo.infotopo import infotopo
