--- infotopo.egg-info/SOURCES.txt
@@ -1,7 +1,0 @@
-README.md
-setup.cfg
-setup.py
-infotopo.egg-info/PKG-INFO
-infotopo.egg-info/SOURCES.txt
-infotopo.egg-info/dependency_links.txt
-infotopo.egg-info/top_level.txt
