+++ infotopo/__init__.py
@@ -1,0 +1,4 @@
+from infotopo.infotopo import Infotopo  # noqa
+from infotopo import io, plot # noqa
+
+__version__ = "0.1.0"
